# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "summary": "Liste des livres",
                "description": "Display a listing of the resource.",
                "operationId": "64ce85f06fc87443a16496268e88cddb",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "summary": "Créer un livre",
                "description": "Store a newly created resource in storage.",
                "operationId": "4114213d13c3b8499c6aec21511edb28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "summary",
                                    "isbn"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nouveau titre unique"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Nouvel auteur unique"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Nouveau sommaire unique"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9780451524910"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livre créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{book}": {
            "get": {
                "summary": "Détails d'un livre",
                "description": "Display the specified resource.",
                "operationId": "13cbe6c41caf25bdab855f6385767c4c",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un livre",
                "description": "Update the specified resource in storage.",
                "operationId": "861d9a24e98c605f9e3471b831a35fae",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nouveau titre"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Nouvel auteur"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Nouveau sommaire"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "9780451524999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livre mis à jour"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Supprimer un livre",
                "description": "Remove the specified resource from storage.",
                "operationId": "c9e60bf997090b51bd0fbedeb013d82f",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID du livre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Livre supprimé"
                    },
                    "404": {
                        "description": "Livre non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "summary": "Inscrire un utilisateur",
                "operationId": "584a102470c054e585330c05f7da4bfc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Thomas"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "thomas@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Connexion d'un utilisateur",
                "operationId": "2526fe2d44c926504aa3663c8f10a15c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "thomas@thomas.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Déconnexion",
                "operationId": "b9c12342da1bfbec9f02fd4e3667ad52",
                "responses": {
                    "204": {
                        "description": "Déconnecté"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                }
            }
        }
    }
}